Add files via upload
I'm not sure if load_pages_to_memory should be inserting into the swapQ (even though her comments say so) because then the queue has those entries as well. Cleaned up some functions in swap and loader mainly to get rid of the segmentations and have a more understandable algorithm for the load to swapQ in loader.